pyproject.toml
[project]
name = "docu-mentor"
version = "0.1.0"
description = "DocuMentor is a smart documentation assistant that helps you quickly query and retrieve relevant information from your documentation. With AI-driven search capabilities, it allows developers and teams to easily access up-to-date and precise details, saving time and effort in navigating complex documents."
authors = [{ name = "rahuljsaliaan", email = "rahuljsaliaan@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13 <3.14"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langsmith (>=0.3.33,<0.4.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "streamlit-chat (>=0.1.1,<0.2.0)",
    "langchain-pinecone (>=0.2.5,<0.3.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "tldextract (>=5.3.0,<6.0.0)",
    "firecrawl-py (>=2.2.0,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "docu_mentor", from = "src" }]

[tool.poetry.scripts]
start = "bin.runner:run"
ui = "src.docu_mentor.ui:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
